+++ Sources/BraintreePayPal/Public/BraintreePayPal/BTPayPalRequest.h
@@ -5,25 +5,6 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- Payment intent.
-
- @note Must be set to sale for immediate payment, authorize to authorize a payment for capture later, or order to create an order. Defaults to authorize. Only works in the checkout (one-time payment) flow.
-
- @see https://developer.paypal.com/docs/integration/direct/payments/capture-payment/ Capture payments later
- @see https://developer.paypal.com/docs/integration/direct/payments/create-process-order/ Create and process orders
-*/
-typedef NS_ENUM(NSInteger, BTPayPalRequestIntent) {
-    /// Authorize
-    BTPayPalRequestIntentAuthorize = 1,
-
-    /// Sale
-    BTPayPalRequestIntentSale,
-
-    /// Order
-    BTPayPalRequestIntentOrder,
-};
-
-/**
  Use this option to specify the PayPal page to display when a user lands on the PayPal site to complete the payment.
 */
 typedef NS_ENUM(NSInteger, BTPayPalRequestLandingPageType) {
@@ -64,21 +45,6 @@ typedef NS_ENUM(NSInteger, BTPayPalRequestUserAction) {
 @interface BTPayPalRequest : NSObject
 
 /**
- Initialize a PayPal request with an amount for a one-time payment.
-
- @param amount Used for a one-time payment. Amount must be greater than or equal to zero, may optionally contain exactly 2 decimal places separated by '.', optional thousands separator ',', and is limited to 7 digits before the decimal point.
- @return A PayPal request.
-*/
-- (instancetype)initWithAmount:(NSString *)amount;
-
-/**
- Used for a one-time payment.
-
- Amount must be greater than or equal to zero, may optionally contain exactly 2 decimal places separated by '.', optional thousands separator ',', and is limited to 7 digits before the decimal point.
-*/
-@property (nonatomic, readonly, strong) NSString *amount;
-
-/**
  Defaults to false. When set to true, the shipping address selector will be displayed.
 */
 @property (nonatomic, getter=isShippingAddressRequired) BOOL shippingAddressRequired;
@@ -89,12 +55,6 @@ typedef NS_ENUM(NSInteger, BTPayPalRequestUserAction) {
  @note Only applies when `shippingAddressOverride` is set.
  */
 @property (nonatomic, getter=isShippingAddressEditable) BOOL shippingAddressEditable;
-
-/**
- Optional: A valid ISO currency code to use for the transaction. Defaults to merchant currency code if not set.
- @note This is only used for one-time payments.
-*/
-@property (nonatomic, nullable, copy) NSString *currencyCode;
 
 /**
  Optional: A locale code to use for the transaction.
@@ -142,11 +102,6 @@ typedef NS_ENUM(NSInteger, BTPayPalRequestUserAction) {
 @property (nonatomic, nullable, copy) NSString *billingAgreementDescription;
 
 /**
- Optional: Payment intent. Only applies when using checkout flow. Defaults to `BTPayPalRequestIntentAuthorize`.
-*/
-@property (nonatomic) BTPayPalRequestIntent intent;
-
-/**
  Optional: Changes the call-to-action in the PayPal flow. This option works for both checkout and vault flows. Defaults to `BTPayPalRequestUserActionDefault`.
 */
 @property (nonatomic) BTPayPalRequestUserAction userAction;
@@ -167,11 +122,6 @@ typedef NS_ENUM(NSInteger, BTPayPalRequestUserAction) {
  Optional: Offers PayPal Credit if the customer qualifies. Defaults to false. Only available with PayPal Checkout and PayPal Billing Agreement.
  */
 @property (nonatomic) BOOL offerCredit;
-
-/**
- Optional: Offers PayPal Pay Later if the customer qualifies. Defaults to false. Only available with PayPal Checkout.
- */
-@property (nonatomic) BOOL offerPayLater;
 
 /**
  Optional: A non-default merchant account to use for tokenization.